u16 -> u8 for MSAA samples

+++ irr/include/IVideoDriver.h
@@ -278,7 +278,7 @@ public:
 	Only works if the driver supports the EVDF_TEXTURE_MULTISAMPLE feature,
 	check via queryFeature.
 	\see addRenderTargetTexture */
-	virtual ITexture *addRenderTargetTextureMs(const core::dimension2d<u32> &size, u16 msaa,
+	virtual ITexture *addRenderTargetTextureMs(const core::dimension2d<u32> &size, u8 msaa,
 			const io::path &name = "rt", const ECOLOR_FORMAT format = ECF_UNKNOWN) = 0;
 
 	//! Adds a new render target texture with 6 sides for a cubemap map to the texture cache.
